requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
jinja2
google-generativeai
python-dotenv
